requirements2.txt
gradio
requests
PyPDF2
sentence-transformers
scikit-learn
streamlit 
PyPDF2 
sentence-transformers 
numpy 
scikit-learn 
openai 
langchain 
langchain-core 
langchain-openai 
yfinance 
google-api-python-client 
google-auth 
google-auth-oauthlib